+++ app/Http/Controllers/ReviewsController.php
@@ -9,12 +9,16 @@ class ReviewsController extends Controller
 {
     public function index()
     {
-        return Review::paginate(10);
+        // return Review::paginate(10);
+
+        return $this->getAuthUser()->reviews()->paginate(10);
     }
 
     public function show(Review $review)
     {
-        return $review;
+        //return $review;
+
+        return $this->authorize('view', $review);
     }
 
     public function store()
@@ -38,6 +42,10 @@ class ReviewsController extends Controller
 
     public function update(Review $review)
     {
+        if ($this->getAuthUser()->cannot('update', $review)) {
+
+            abort(403);
+        }
 
         $attributes = request()->validate([
             'name' => 'min:3|max:10'
@@ -55,6 +63,10 @@ class ReviewsController extends Controller
     public function destroy(Review $review)
     {
 
+        if ($this->getAuthUser()->cannot('delete', $review)) {
+
+            abort(403);
+        }
         $review->delete();
 
         return back()->with('success', 'Review Deleted');
